requirements.txt
asgiref==3.2.7
coverage==5.1
Django==3.0.6
django-extensions==2.2.9
djangorestframework==3.11.0
djangorestframework-gis==0.15
isodate==0.6.0
psycopg2-binary==2.8.5
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
